cephfs: encode the subvolume metrics count as a fixed-width type
encode() was called with the result of std::vector::size(), i.e. size_t.
There is no exact ceph::encode() overload for size_t on platforms where
uint64_t is unsigned long long rather than unsigned long, so the call is
ambiguous there.

Encode uint64_t explicitly.  That is byte for byte what LP64 platforms were
already putting on the wire, so the encoding is unchanged.

+++ src/include/cephfs/metrics/Types.h
@@ -754,7 +754,7 @@ struct SubvolumeMetricsPayload : public ClientMetricPayloadBase {
   void encode(bufferlist &bl) const {
     using ceph::encode;
     ENCODE_START(1, 1, bl);
-    encode(subvolume_metrics.size(), bl);
+    encode(static_cast<uint64_t>(subvolume_metrics.size()), bl);
     for(auto const& m : subvolume_metrics) {
       m.encode(bl);
     }
@@ -764,7 +764,7 @@ struct SubvolumeMetricsPayload : public ClientMetricPayloadBase {
   void decode(bufferlist::const_iterator &iter) {
     using ceph::decode;
     DECODE_START(1, iter) ;
-    size_t size = 0;
+    uint64_t size = 0;
     decode(size, iter);
     subvolume_metrics.clear();
     subvolume_metrics.reserve(size);
